chatbot-backend/requirements.txt
fastapi
uvicorn
cohere
openai
qdrant-client
python-dotenv